Build etch a sketch
Create a webpage with a 16x16 grid of square divs. Set up a “hover”
effect so that the grid divs change color when your mouse passes over
them, leaving a (pixelated) trail through your grid like a pen would.

Add a button on the top of the screen that will send the user a popup
asking for the number of squares per side for the new grid.